+++ example/example.dart
@@ -38,18 +38,38 @@ Future<void> main() async {
 
   // Print title of the Bar instance associated with Foo.
   print(dependency.foo.bar.title);
+
+  print(dependency.blocs.foo().time);
+  print(dependency.blocs.foo().time);
 }
 
 /// Class representing a dependency.
 class Dependency {
   final Foo foo;
+  final Blocs blocs;
 
   /// Factory method to create Dependency instances.
   static Dependency onInject(DepsCallback deps) {
-    return Dependency(foo: deps<Foo>());
+    final factory = deps<SyringeFactory>();
+
+    return Dependency(foo: deps<Foo>(), blocs: Blocs(factory: factory));
   }
 
-  const Dependency({required this.foo});
+  const Dependency({required this.foo, required this.blocs});
+}
+
+final class Blocs {
+  final SyringeFactory _factory;
+
+  Foo foo() {
+    return _factory.create<Foo>(
+        deps: [Bar],
+        onFactory: (factory) {
+          return Foo(bar: factory<Bar>());
+        });
+  }
+
+  Blocs({required SyringeFactory factory}) : _factory = factory;
 }
 
 /// Class representing a Bar instance.
@@ -62,8 +82,11 @@ class Bar {
 /// Class representing a Foo instance.
 class Foo {
   final Bar bar;
+  final DateTime time;
 
-  const Foo({required this.bar});
+  Foo({
+    required this.bar,
+  }) : time = DateTime.now();
 }
 
 /// Module for creating Bar instance.
@@ -78,5 +101,5 @@ final class BarModule extends Module<Bar> {
   }
 
   @override
-  bool get export => false;
+  bool get export => true;
 }